added authentication steps to basic crud
Former-commit-id: 6e2a0d44941a9208f19d366c6579940fe69c4466 [formerly 54d5e1d81253e040ea03b6fec979365f559c1f70]
Former-commit-id: fc895ed9c93e392edf9b2853c364e9a205a7d1e3

+++ api/src/projects/deleteProject.resolver.ts
@@ -1,9 +1,13 @@
-import { Resolver, ArgsType, Field, Args, Mutation } from 'type-graphql';
-import { logger } from '@typegoose/typegoose/lib/logSettings';
+import { Resolver, ArgsType, Field, Args, Mutation, Ctx } from 'type-graphql';
 import { projectIndexName } from '../elastic/settings';
 import { elasticClient } from '../elastic/init';
 import { ObjectId } from 'mongodb';
 import { ProjectModel } from '../schema/project';
+import { getLogger } from 'log4js';
+import { GraphQLContext } from '../utils/context';
+import { verifyLoggedIn } from '../auth/checkAuth';
+import { checkAccess } from '../utils/checkAccess';
+import { AccessLevel } from '../schema/access';
 
 @ArgsType()
 class DeleteProjectArgs {
@@ -11,11 +15,22 @@ class DeleteProjectArgs {
   id: ObjectId;
 }
 
+const logger = getLogger();
+
 @Resolver()
 class DeleteProjectResolver {
   @Mutation(_returns => String)
-  async deleteProject(@Args() args: DeleteProjectArgs): Promise<string> {
-    const id = new ObjectId();
+  async deleteProject(@Args() args: DeleteProjectArgs, @Ctx() ctx: GraphQLContext): Promise<string> {
+    if (!verifyLoggedIn(ctx) || !ctx.auth) {
+      throw new Error('user not logged in');
+    }
+    const project = await ProjectModel.findById(args.id);
+    if (!project) {
+      throw new Error(`cannot find project with id ${args.id.toHexString()}`);
+    }
+    if (!checkAccess(new ObjectId(ctx.auth.id), project.access, AccessLevel.admin)) {
+      throw new Error('user does not have admin access to project');
+    }
     const deleteElasticResult = await elasticClient.delete({
       index: projectIndexName,
       id: args.id.toHexString()
@@ -24,7 +39,7 @@ class DeleteProjectResolver {
     await ProjectModel.deleteOne({
       _id: args.id
     });
-    return `deleted project with id: ${id}`;
+    return `deleted project with id: ${args.id.toHexString()}`;
   }
 }
